[chore] Add integration test
This change adds an integration test that will create a test repo, a
test remote, spin up an HTTP server with basic auth and verify that
we can fetch from the remote with valid credentials but not without.

This serves both as an example as well as a test of how this can
be used.

Signed-off-by: Dominik Schulz <dominik.schulz@gauner.org>